% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipest.R
\name{EstimateOneTip}
\alias{EstimateOneTip}
\title{Infer state at one tip}
\usage{
EstimateOneTip(tip_to_reconstruct, existing_tips, rate_matrix, phy)
}
\arguments{
\item{tip_to_reconstruct}{Label of the tip to reconstruct}

\item{existing_tips}{Vector of tip states, with labels equal to tips on the tree}

\item{rate_matrix}{Rate matrix to pass to geiger for reconstruction}

\item{phy}{A phylo object (your tree)}
}
\value{
A list with $best (the state of the best tip) and $all, a data.frame of all tip states and their AICc values
}
\description{
This function estimates the best tip state (given AICc) but also gives you access to the AICc values for all the possible tip states.
}
\details{
Note that this gives the best solution given the observed tips and ONLY this tip. It does not compare with other tips to reconstruct to get the best global solution.
}
